requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.15.0
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.4.2
pytest==8.2.0

